+++ Content.Client/_Arcane/Leash/LeashOverlaySystem.cs
@@ -1,16 +1,18 @@
+using Content.Shared._Arcane.Leash;
 using Robust.Client.Graphics;
 
 namespace Content.Client._Arcane.Leash;
 
-public sealed class LeashSystem : EntitySystem
+public sealed class LeashSystem : SharedLeashSystem
 {
     [Dependency] private readonly IOverlayManager _overlayManager = default!;
 
-    private LeashOverlay _overlay = default!;
+    private LeashOverlay? _overlay;
 
     public override void Initialize()
     {
         base.Initialize();
+
         _overlay = new LeashOverlay(EntityManager);
         _overlayManager.AddOverlay(_overlay);
     }
@@ -18,6 +20,11 @@ public sealed class LeashSystem : EntitySystem
     public override void Shutdown()
     {
         base.Shutdown();
-        _overlayManager.RemoveOverlay(_overlay);
+
+        if (_overlay != null)
+        {
+            _overlayManager.RemoveOverlay(_overlay);
+            _overlay = null;
+        }
     }
 }
